requirements.txt
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-mssql==1.8
django-mssql-backend==2.8.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
drf-spectacular==0.9.3
drf-writable-nested==0.5.1
drf-yasg==1.17.1
idna==2.8
importlib-metadata==1.6.0
inflection==0.4.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==20.3
Pillow==7.1.2
pycodestyle==2.5.0
pycrypto==2.6.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.16.0
python-tds==1.10.0
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
zipp==3.1.0
